Close comparison modal and clear selected checkboxes on Search page unmount
