+++ src/test/__tests__/views/Search.test.tsx
@@ -1,6 +1,8 @@
 import '@src/test/__mocks__/common/helpers/pageScrolling.helper.mock';
 import { render, screen } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
+import { setInitialGlobalState } from '@src/test/__mocks__/store';
+import { useSearchStore, useUIStore } from '@src/store';
 import { Search } from '@views';
 
 jest.mock('@common/constants/build.constants', () => ({ IS_EMBEDDED_MODE: false }));
@@ -20,5 +22,38 @@ describe('Search', () => {
 
   test('renders child ItemSearch component', () => {
     expect(screen.getByTestId('id-search')).toBeInTheDocument();
+  });
+
+  describe('component unmount', () => {
+    test('clears fullDisplayComponentType and selectedInstances on unmount', () => {
+      const mockResetFullDisplayComponentType = jest.fn();
+      const mockResetSelectedInstances = jest.fn();
+
+      setInitialGlobalState([
+        {
+          store: useUIStore,
+          state: {
+            resetFullDisplayComponentType: mockResetFullDisplayComponentType,
+          },
+        },
+        {
+          store: useSearchStore,
+          state: {
+            resetSelectedInstances: mockResetSelectedInstances,
+          },
+        },
+      ]);
+
+      const { unmount } = render(
+        <BrowserRouter>
+          <Search />
+        </BrowserRouter>,
+      );
+
+      unmount();
+
+      expect(mockResetFullDisplayComponentType).toHaveBeenCalled();
+      expect(mockResetSelectedInstances).toHaveBeenCalled();
+    });
   });
 });
